fix: move grpc objects to same file
